requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.96.0
anyio==4.13.0
asyncpg==0.31.0
attrs==26.1.0
certifi==2026.2.25
discord.py==2.7.1
distro==1.9.0
docstring_parser==0.18.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.14.0
multidict==6.7.1
propcache==0.4.1
pydantic==2.13.2
pydantic_core==2.46.2
python-dotenv==1.2.2
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
yarl==1.23.0
